requirements.txt
pyyaml>=6.0
gitpython>=3.1.31
requests>=2.31.0
python-dotenv>=1.0.0
ipython>=8.0.0
langchain>=1.0.0
langchain-google-genai>=0.0.13